# backend/routers/users.py
"""Users router."""

from typing import List, Optional
from uuid import UUID

from fastapi import APIRouter

from dependencies import CogniterUser, UserServiceDep
from exceptions import AuthorizationError
from models import UserRole
from schemas import UserRead
from schemas.user import UserRoleUpdate

router = APIRouter(
    prefix="/users",
    tags=["users"],
)


@router.get("/", response_model=List[UserRead])
async def list_users(
    current_user: CogniterUser,  # Only Cogniters can list users
    user_service: UserServiceDep,
    search: Optional[str] = None,
    role: Optional[UserRole] = None,
    limit: int = 50,
):
    """
    List users with optional filtering.

    - **search**: Optional search string for name or email (case-insensitive)
    - **role**: Optional filter by role (Cogniter, Client + Financials, or Client)
    - **limit**: Maximum number of results to return (default 50, max 200)

    Only Cogniters can access this endpoint.
    """
    # Cap limit to prevent abuse
    effective_limit = min(limit, 200) if limit > 0 else 50
    return user_service.search_users(search=search, role=role, limit=effective_limit)


@router.patch("/{user_id}/role", response_model=UserRead)
async def update_user_role(
    user_id: UUID,
    role_data: UserRoleUpdate,
    current_user: CogniterUser,  # Only Cogniters can update roles
    user_service: UserServiceDep,
):
    """
    Update a user's role.

    - Only Cogniters can update user roles
    - Cannot promote anyone to Cogniter role (security)
    - Can only change between CLIENT and CLIENT_FINANCIALS roles

    Use this to upgrade a client to Client + Financials for financial data access.
    """
    # Security: Cannot promote to Cogniter
    if role_data.role == UserRole.COGNITER:
        raise AuthorizationError("Cannot promote users to Cogniter role")

    # ResourceNotFoundError -> 404 via global handler
    return user_service.update_role(user_id, role_data.role)